Move all slack code into module
