+++ lib/slack/unfurl/index.js
@@ -4,13 +4,13 @@
 
 // likely need different regular expressions based on what we're trying to parse
 
-const githubUrl = require('./github-url');
-const unfurlComment = require('./unfurl/comment');
-const unfurlBlob = require('./unfurl/blob');
-const unfurlAccount = require('./unfurl/account');
-const unfurlIssue = require('./unfurl/issue');
-const unfurlPull = require('./unfurl/pull');
-const unfurlRepo = require('./unfurl/repo');
+const githubUrl = require('../../github-url');
+const unfurlComment = require('./comment');
+const unfurlBlob = require('./blob');
+const unfurlAccount = require('./account');
+const unfurlIssue = require('./issue');
+const unfurlPull = require('./pull');
+const unfurlRepo = require('./repo');
 
 module.exports = async (github, url) => {
   const params = githubUrl(url);
