+++ lib/slack/unfurl/issue.js
@@ -3,7 +3,7 @@ const ISSUE_OPEN_GREEN = '#36a64f';
 
 const moment = require('moment');
 
-const { arrayToFormattedString } = require('../helpers');
+const { arrayToFormattedString } = require('../../helpers');
 
 module.exports = function unfurlIssue(issue) {
   return {
