+++ lib/slack/unfurl/pull.js
@@ -4,7 +4,7 @@ const {
   arrayToFormattedString,
   getHexColorbyState,
   getFormattedState,
-} = require('../helpers');
+} = require('../../helpers');
 
 module.exports = function unfurlPull(pull, params) {
   return {
